requirements.txt
grpcio-tools ~= 1.30
flask ~= 1.1
Jinja2 ~= 2.11
